pyproject.toml
[tool.poetry]
name = "intermediatedjango"
version = "0.1.0"
description = ""
authors = ["Renan Nunes Bittencourt <programmerrnb@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
whitenoise = "^6.8.2"
gunicorn = "^23.0.0"
django-stdimage = "^6.0.2"
psycopg2-binary = "^2.9.10"
django-bootstrap5 = "^24.3"
pytest-mock = "^3.14.0"
pillow = "^11.0.0"
dj-database-url = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.8.3"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
python_files = ["test*.py","*_test.py"]
addopts = "--strict-markers"

[tool.ruff]
line-length=120
exclude = ['./core/migrations/']

[tool.ruff.lint]
select = ['I','F','E','W','PL','PT']
ignore = ["PLR2004"]

[tool.ruff.format]
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'python manage.py runserver'
post_test = 'coverage html'
pre_test = 'task lint'
test = 'pytest . --cov=. -vv'
mkmigrations = 'python manage.py makemigrations'
migrate = 'python manage.py migrate'
shell = 'python manage.py shell'
static = 'python manage.py collectstatic'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
